backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
motor==3.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.1.0
python-dotenv==1.0.0
pymongo==4.3.3
typing-extensions==4.5.0
starlette==0.26.1